+++ database/migrations/2020_06_24_013227_create_permission_user_table.php
@@ -14,8 +14,8 @@ class CreatePermissionUserTable extends Migration
     public function up()
     {
         Schema::create('permission_user', function (Blueprint $table) {
-            $table->bigInteger('user_id')->unsigned();
-            $table->bigInteger('permission_id')->unsigned();
+            $table->integer('user_id')->unsigned();
+            $table->integer('permission_id')->unsigned();
         
          //FOREIGN KEY CONSTRAINTS
             $table->foreign('user_id')
